requirements.txt
streamlit
torch
torchvision
Pillow
facenet-pytorch